backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# Redis & Caching
redis==5.0.1

# AI/ML - CLIP Embeddings
torch==2.2.0
transformers==4.37.2
sentence-transformers==2.3.1
Pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.0

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
